refactor: divide tag components
